Rearrange taskbar windows to match the switching order.
This makes it so the windows in the taskbar exactly match the ordering in
the alt-tab quickswitch popup.

The benefit of this for me is that recently used windows are always at the
right (can be changed to be at the left), and I can see where they all are,
and work out in advance how long to hold down alt-tab for, to get to the
window.  Without this, you need to press alt-tab, release tab, search for
the window, then hold tab to get to it.  Also usually I have so many windows
the vertical list is truncated off the bottom of the screen.

This window switching method of course conflicts with the other recently
added option of manually moving the focused window's position
(i.e. Alt+Shift+1), and currently breaks drag and drop (possible to fix that
issue with more work), but in my opinion this method makes those methods of
managing windows unnessecary.

This is a rewrite of https://sourceforge.net/p/icewm/patches/201/.  I've been
using this since then and can't live without it, but have only just started
using this rewrite, so it hasn't had much testing.